update after moved
